+++ UWUVCI AIO WPF/Models/TKeys.cs
@@ -1,7 +1,7 @@
 ﻿using GameBaseClassLibrary;
 using System;
 
-namespace UWUVCI_AIO_WPF.Classes
+namespace UWUVCI_AIO_WPF.Models
 {
     [Serializable]
     public class TKeys
